requirements.txt
Flask>=3.0,<4.0
SQLAlchemy>=2.0.49,<3.0
alembic>=1.13,<2.0
PyMySQL>=1.1,<2.0
python-dotenv>=1.0,<2.0
